+++ CHANGES.md
@@ -15,6 +15,7 @@
  * Use [Z3-TurnKey](https://github.com/tudo-aqua/z3-turnkey) instead of a
    bespoke Z3 build, see #219
  * Use Z3 version 4.8.7.1, see #219
+ * Re-stabilized tests on recursive operators
 
 ## 0.7.2 [RELEASE]